requirements.txt
streamlit
openai
python-dotenv
pandas
matplotlib
faster-whisper
pydub
PyPDF2
numpy
reportlab
jinja2
librosa
soundfile